// hornet-js-core/src/result/hornet-result-interface.ts
export interface Options {
    data?: any;
    filename? : string;
    encoding? : string;
}

export interface OptionsFiles extends Options {
    size?: number;
}

export interface OptionsPDF extends OptionsFiles {
    definition: {},
    fonts?:{}
}